requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-CORS
Flask-JWT-Extended
pytest-flask